refactor: refactored prompt converting files for llama, mistral and nova and extracted helper functions to separate files for much neater code
